Initial commit of Manifest Version 3 compliant version
The majority of testing was done in Chromium Version 113.0.5672.64 (Official Build) (64-bit) on Windows 10 (https://chromium.woolyss.com)

Note that in order for chrome.idle detection to work we need to to persist the Service Worker.
This was accomplished with a "hack", as Service Workers are meant to be ephemeral & self-terminating.
For the "hack" details, see https://stackoverflow.com/questions/66618136/persistent-service-worker-in-chrome-extension#:~:text=%22Persistent%22%20service%20worker%20with%20offscreen%20API

There is also a significant amount of file reorganisation in this version.