Fix "Uncaught TypeError: Cannot read properties of null (reading '1')"
When the option values aren't supplied as part of the UA string, i.e. in most deployments, the variable suppliedByUA will be null.
We need to check that this variable is not null before testing its values are lower/higher than the min/max values, otherwise the above TypeError will result.

This fix is connected with the previous commit, 560b78e.